enclave: Ported to using new phactory crate
