+++ standalone/pruntime/app/src/main.rs
@@ -252,77 +252,6 @@ fn ocall_get_update_info(
     }
 }
 
-#[no_mangle]
-pub extern "C"
-fn ocall_dump_state(
-    _output_ptr : *mut u8,
-    _output_len_ptr: *mut usize,
-    _output_buf_len: usize
-) -> sgx_status_t {
-    // TODO:
-
-    sgx_status_t::SGX_SUCCESS
-}
-
-#[no_mangle]
-pub extern "C"
-fn ocall_save_persistent_data(
-    input_ptr: *const u8,
-    input_len: usize
-) -> sgx_status_t {
-    let input_slice = unsafe { std::slice::from_raw_parts(input_ptr, input_len) };
-    debug!("Sealed data {:}: {:?}", input_len, hex::encode(input_slice));
-
-    let executable = env::current_exe().unwrap();
-    let path = executable.parent().unwrap();
-    let state_path: path::PathBuf = path.join(*ENCLAVE_STATE_FILE_PATH).join(ENCLAVE_STATE_FILE);
-    info!("Save seal data to {}", state_path.as_path().to_str().unwrap());
-
-    fs::write(state_path.as_path().to_str().unwrap(), input_slice)
-        .expect("Failed to write persistent data");
-
-    sgx_status_t::SGX_SUCCESS
-}
-
-#[no_mangle]
-pub extern "C"
-fn ocall_load_persistent_data(
-    output_ptr : *mut u8,
-    output_len_ptr: *mut usize,
-    output_buf_len: usize
-) -> sgx_status_t {
-    let executable = env::current_exe().unwrap();
-    let path = executable.parent().unwrap();
-    let state_path: path::PathBuf = path.join(*ENCLAVE_STATE_FILE_PATH).join(ENCLAVE_STATE_FILE);
-
-    let state = match fs::read(state_path.as_path().to_str().unwrap()) {
-        Ok(data) => data,
-        _ => Vec::<u8>::new()
-    };
-    let state_len = state.len();
-
-    if state_len == 0 {
-        return sgx_status_t::SGX_SUCCESS
-    }
-
-    info!("Loaded sealed data {:}: {:?}", state_len, state);
-
-    unsafe {
-        if state_len <= output_buf_len {
-            std::ptr::copy_nonoverlapping(state.as_ptr(),
-                                          output_ptr,
-                                          state_len);
-        } else {
-            panic!("State too long. Buffer overflow.");
-        }
-        std::ptr::copy_nonoverlapping(&state_len as *const usize,
-                                      output_len_ptr,
-                                      std::mem::size_of_val(&state_len));
-    }
-
-    sgx_status_t::SGX_SUCCESS
-}
-
 fn init_enclave() -> SgxResult<SgxEnclave> {
     let mut launch_token: sgx_launch_token_t = [0; 1024];
     let mut launch_token_updated: i32 = 0;
