+++ standalone/pruntime/enclave/Enclave.edl
@@ -36,10 +36,6 @@ enclave {
         );
 
         public sgx_status_t ecall_bench_run(uint32_t index);
-
-        public sgx_status_t ecall_set_state(
-            [in, size = input_len] uint8_t *input_ptr, size_t input_len
-        );
     };
 
     untrusted {
@@ -75,20 +71,6 @@ enclave {
             [in] sgx_platform_info_t * platformBlob,
             int32_t enclaveTrusted,
             [out] sgx_update_info_bit_t * update_info
-        );
-
-        sgx_status_t ocall_dump_state(
-            [out, size = output_buf_len] uint8_t *output_ptr,
-            [out] size_t *output_len_ptr, size_t output_buf_len
-        );
-
-        sgx_status_t ocall_save_persistent_data(
-            [in, size = input_len] uint8_t *input_ptr, size_t input_len
-        );
-
-        sgx_status_t ocall_load_persistent_data(
-            [out, size = output_buf_len] uint8_t *output_ptr,
-            [out] size_t *output_len_ptr, size_t output_buf_len
         );
     };
 };
